Revisions to edge-to-edge map controls. Pending maplibre updates
